Where to begin ... Rearchitected the whole model/route relationship, abstracting out a common "resource" class and then subclassing it for the various types of objects we're dealing with. Each subclass, when necessary, overrides its parent functions for things like setting attributes, modifying the bundle, where to store itself, etc.
Also added in capability to have client consumers/serializers, when the API needs to send or receive data in a different form. Currently Popcorn is the only client.

Other little changes here and there ... just a generic route for lists and another for details, calling the proper method function of the resource when needed.

Finally, added a 404 handler when an object isn't found.