Some big changes
1) Added in an authorization level. Authorization now works like this:

a) Every object has a "dc:coverage" attribute. If this is "public," that object will be visible via the api regardless of authentication status. If this is "BYU," that object will be visible to any authenticated user, regardless of role. If that object is "private," it will be visible only to the owner, to superusers, and to those who pass the 'rights' test (see next item). Eventually, for course collections and videos that are members of course collections, private status will be visible to students enrolled in those courses as well.
b) Every object has a "dc:rights" attribute that has two child attributes, "read" and "write." Both child attributes are lists of netIDs. If someone's netID shows up in the "read" list, he/she can see that object when authenticated, regardless of role or enrollment. Nothing has been implemented for "write" yet, but that will eventually be used to control POST, PUT, PATCH, and DELETE (i.e. you must have that write right, be the owner, or be a superuser to perform one of those.)
c) Default coverage for collection is "BYU," for video is "private." Default rights is empty.
d) The collection api will also filter out any videos that a requestor can't see ...so someone may be able to get collection data with an empty set of videos, while others may have videos in there.

2) Added in the ability to manually authenticate as a test user. Those will be distributed (along with methods for doing so) to developers.

3) fixed some bugs in the logout.